Refactor keycode
